refactor state to use "immer"
