+++ package.json
@@ -13,11 +13,13 @@
   "dependencies": {
     "eslint": "8.35.0",
     "eslint-config-next": "13.2.3",
+    "immer": "^9.0.21",
     "next": "13.2.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "styled-components": "^5.3.8",
     "uid": "^2.0.1",
+    "use-immer": "^0.9.0",
     "zustand": "^4.3.6"
   },
   "devDependencies": {
